Promote v2 rewrite to src root and move data to canvas coordinates
Remove the legacy renderer (old/ and the non-v2 src files), promote
src/v2/* into src/, and point the rollup entry at src/main.ts.

Rewrite the six deployment YAML files from a centre-origin coordinate
system to canvas/top-left origin (offset by [30, 22]), matching the
rewritten getCoordinates which no longer translates by a CENTER anchor.
isCenterObjective now compares against the canvas centre instead of
[0, 0].
