+++ src/building-coordinates.test.ts
