+++ src/building-coordinates.ts
