+++ static/data/deployment/crucible_of_battle.yml
@@ -3,19 +3,19 @@ home_edge: short
 
 attacker:
   deployment_zone:
-    - [30, -22]
-    - [30, 22]
-    - [0, -22]
+    - [60, 0]
+    - [60, 44]
+    - [30, 0]
 
 defender:
   deployment_zone:
-    - [0, 22]
-    - [-30, 22]
-    - [-30, -22]
+    - [30, 44]
+    - [0, 44]
+    - [0, 0]
 
 objectives:
-  - [0, 0]
-  - [10, -14]
-  - [16, 12]
-  - [-10, 14]
-  - [-16, -12]
+  - [30, 22]
+  - [40, 8]
+  - [46, 34]
+  - [20, 36]
+  - [14, 10]
