+++ static/data/deployment/dawn_of_war.yml
@@ -3,21 +3,21 @@ home_edge: long
 
 attacker:
   deployment_zone:
-    - [-30, -22]
-    - [-30, -10]
-    - [30, -10]
-    - [30, -22]
+    - [0, 0]
+    - [0, 12]
+    - [60, 12]
+    - [60, 0]
 
 defender:
   deployment_zone:
-    - [-30, 10]
-    - [-30, 22]
-    - [30, 22]
-    - [30, 10]
+    - [0, 32]
+    - [0, 44]
+    - [60, 44]
+    - [60, 32]
 
 objectives:
-  - [0, 0]
-  - [-20, 0]
-  - [20, 0]
-  - [0, 16]
-  - [0, -16]
+  - [30, 22]
+  - [10, 22]
+  - [50, 22]
+  - [30, 38]
+  - [30, 6]
