+++ static/data/deployment/hammer_and_anvil.yml
@@ -3,21 +3,21 @@ home_edge: short
 
 attacker:
   deployment_zone:
-    - [30, -22]
-    - [12, -22]
-    - [12, 22]
-    - [30, 22]
+    - [60, 0]
+    - [42, 0]
+    - [42, 44]
+    - [60, 44]
 
 defender:
   deployment_zone:
-    - [-12, -22]
-    - [-30, -22]
-    - [-30, 22]
-    - [-12, 22]
+    - [18, 0]
+    - [0, 0]
+    - [0, 44]
+    - [18, 44]
 
 objectives:
-  - [0, 0]
-  - [0, 16]
-  - [0, -16]
-  - [20, 0]
-  - [-20, 0]
+  - [30, 22]
+  - [30, 38]
+  - [30, 6]
+  - [50, 22]
+  - [10, 22]
