+++ static/data/deployment/search_and_destroy.yml
@@ -4,22 +4,22 @@ home_edge: long
 attacker:
   mask_center: 9
   deployment_zone:
-    - [30, -22]
+    - [60, 0]
+    - [60, 22]
+    - [30, 22]
     - [30, 0]
-    - [0, 0]
-    - [0, -22]
 
 defender:
   mask_center: 9
   deployment_zone:
-    - [0, 0]
+    - [30, 22]
+    - [30, 44]
+    - [0, 44]
     - [0, 22]
-    - [-30, 22]
-    - [-30, 0]
 
 objectives:
-  - [0, 0]
-  - [-16, -12]
-  - [-16, 12]
-  - [16, -12]
-  - [16, 12]
+  - [30, 22]
+  - [14, 10]
+  - [14, 34]
+  - [46, 10]
+  - [46, 34]
