+++ static/data/deployment/sweeping_engagement.yml
@@ -3,25 +3,25 @@ home_edge: long
 
 attacker:
   deployment_zone:
-    - [30, 22]
-    - [-30, 22]
-    - [-30, 8]
-    - [0, 8]
-    - [0, 14]
-    - [30, 14]
+    - [60, 44]
+    - [0, 44]
+    - [0, 30]
+    - [30, 30]
+    - [30, 36]
+    - [60, 36]
 
 defender:
   deployment_zone:
-    - [-30, -22]
-    - [30, -22]
-    - [30, -8]
-    - [0, -8]
-    - [0, -14]
-    - [-30, -14]
+    - [0, 0]
+    - [60, 0]
+    - [60, 14]
+    - [30, 14]
+    - [30, 8]
+    - [0, 8]
 
 objectives:
-  - [0, 0]
-  - [-20, -4]
-  - [-12, 16]
-  - [12, -16]
-  - [20, 4]
+  - [30, 22]
+  - [10, 18]
+  - [18, 38]
+  - [42, 6]
+  - [50, 26]
