+++ static/data/deployment/tipping_point.yml
@@ -3,26 +3,26 @@ home_edge: short
 
 attacker:
   deployment_zone:
-    - [-30, 22]
-    - [-30, -22]
-    - [-18, -22]
-    - [-18, 0]
-    - [-10, 0]
-    - [-10, 10]
-    - [-10, 22]
+    - [0, 44]
+    - [0, 0]
+    - [12, 0]
+    - [12, 22]
+    - [20, 22]
+    - [20, 32]
+    - [20, 44]
 
 defender:
   deployment_zone:
-    - [30, -22]
-    - [30, 22]
-    - [18, 22]
-    - [18, 0]
-    - [10, 0]
-    - [10, -22]
+    - [60, 0]
+    - [60, 44]
+    - [48, 44]
+    - [48, 22]
+    - [40, 22]
+    - [40, 0]
 
 objectives:
-  - [0, 0]
-  - [-8, -14]
-  - [-16, 12]
-  - [8, 14]
-  - [16, -12]
+  - [30, 22]
+  - [22, 8]
+  - [14, 34]
+  - [38, 36]
+  - [46, 10]
